+++ src/main/java/wooteco/subway/dto/request/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.request;
 
 import wooteco.subway.domain.Station;
 
@@ -6,7 +6,7 @@ public class StationRequest {
 
     private String name;
 
-    public StationRequest() {
+    private StationRequest() {
     }
 
     public StationRequest(String name) {
